style: HomeLayout을 Home으로 병합
