--- src/home/components/HomeLayout/HomeLayout.style.ts
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-export const StyledLayout = styled.div`
-  width: 100vw;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding-bottom: 12.5rem;
-  padding-right: 4.25rem;
-  padding-left: 4.25rem;
-  padding-top: 4.69rem;
-`;
