+++ src/home/pages/Home/Home.style.ts
@@ -1,6 +1,17 @@
 import styled from 'styled-components';
 
+export const StyledContainer = styled.div`
+  min-height: 100vh;
+  position: relative;
+`;
+
 export const StyledComponentContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
+
+export const StyledComponentInnerContainer = styled.div`
   display: flex;
   gap: 1.1875rem;
   margin-top: 1.5rem;
