+++ src/router.tsx
@@ -9,11 +9,6 @@ const Search = lazy(() =>
     default: Search,
   }))
 );
-const HomeLayout = lazy(() =>
-  import('./home/components/HomeLayout/HomeLayout').then(({ HomeLayout }) => ({
-    default: HomeLayout,
-  }))
-);
 const DrawerLayout = lazy(() =>
   import('./drawer/components/Layout/Layout').then(({ Layout }) => ({
     default: Layout,
@@ -105,11 +100,9 @@ export const Router = () => {
       }
     >
       <Routes>
-        <Route path="/" element={<HomeLayout />}>
-          <Route path="" element={<Home />}></Route>
-          <Route path="/login" element={<Login />}></Route>
-          <Route path="/signup" element={<Signup />}></Route>
-        </Route>
+        <Route path="/" element={<Home />}></Route>
+        <Route path="/login" element={<Login />}></Route>
+        <Route path="/signup" element={<Signup />}></Route>
         <Route path="/mypage" element={<Mypage />}></Route>
         <Route path="/drawer" element={<DrawerLayout />}>
           <Route index element={<Navigate to="rankings" replace />}></Route>
